svg: Skip rendering empty words
Today, when smartWordToGlyphs is given words like "-" or "/", it can
return an empty word. And the SVG functions render empty words with a
zero-length word line, resulting in showing just a dot.

This patch fixes the problem by making the SVG rendering skip empty
words.

While we could return an error, in practice just skipping the word is
more user friendly.